v1.2.1 release candidate (2) (#119)
* Update README.md

* Added warning comments to evaluate warnnings

* Resolved warnings

* fix plot not showing up on compare results view

* updated Tutorial 2b

* Updated Canflood Zip

* Update curvePlot.py

* tweaks and requirement changes

* clean up results tests

* add test_results_02_runcompare test case

* reduce some logging

* fix the comparison tool

* clean up test case

* modcom . fix relative filepath catch

* clean up log

* Update pytest.ini

* clean up plt_window

* results.dialog fix run_plotRisk kwargs

* implement radioButton_s_saveToFile

* Updated Plt window logic for showing plots

* remove a link

---------

+++ canflood/results/reporter.py
@@ -454,8 +454,7 @@ class ReportGenerator(RiskPlotr, Qcoms):
         if df is not None:
             #format table
             df.columns = map(str.upper, df.columns)
-            df.iloc[:, 1] = df.iloc[:, 1].map(lambda impact_val: self.impactFmtFunc(impact_val))
-            
+            df.iloc[:, 1] = df.iloc[:, 1].map(lambda impact_val: self.impactFmtFunc(impact_val)).astype(float)            
             #convert to layer
             df_layer = self.vlay_new_df2(df, layname='event_summary_table',
                                        logger=log)
